Move Layouts children to data
The zip file installation extracts it's content directly into the data
directory. This is unacceptable, the data dir is read-only.

Instead, install files at the beginning as if they were extracted
from the zip file. Also, disable buttons to Install and Uninstall,
these are just not sane -- installation and uninstall is handled
by external tools.